added a toy test case.
